Debut Commit!
The current state of the program includes lots of test code, and a rough
idea of the general structure of the classes. The output produces a
green screen, with a mock version of The Add Button. 
- The button is draggable.
- Clicking it creates an Entry Bubble, at some location based on the
current number of Entries in the project.
- Dragging the button over the entries deletes them.
- Clicking the Entries deletes them.
- Entries shift over in response to deleted entries. (If you make a
bunch of entries, you can get a PacMan effect.)

Still a long ways to go.