This is the TRIFORCE* COMMIT!
Hacked together a release. Now, I'm not exactly sure what it looked like on the last commit, but I definitely see a hell of a lot of red in the file explorer, so this comment probably won't cover everything. (I'm just gonna read off the red.)

- Created the ScreenController, along with the abstract Screen class (and a few subclasses). See comments for details!
- Implemented a basic version of the Overview and Startup Menu, as well as added placeholders for the other screens.
- Mustered some implementation of the EntryCell interface.
- Implemented a popup for adding Entries that showcases how hard it actually ISN'T.
- Realized I spend too much time writing commit comments.
- Red, green(ish)... very BADLY need to refactor.
- *Tapped into my ability to do a lot, very quickly.

Here's to the future.